requirements.txt
Pillow==11.0.0
streamlit==1.40.2
tensorflow==2.18.0